refactor(app): unify chat-target selection into one shared hook
The sidebar, home composer, and settings each reimplemented the same selection
orchestration (load persisted selection, build targets, resolve, repair, persist
on select, couple the LLM default), which is why the agent-default bug appeared in
more than one place. Consolidate it:

- commitChatTargetSelection: one shared change-provider side effect (persist the
  selection and, for an LLM target, update the default-provider id), used by all
  three surfaces so the two stores never drift.
- useChatTargetSelection: one hook owning the selection lifecycle for the dropdown
  surfaces; useChatRefs and AgentCommandHome now delegate to it.
- Settings routes its radio actions through the shared side effect while keeping
  its own effectiveTarget display.

No user-visible change; single source of truth so the persistence bug cannot recur
as new surfaces are added.
